Version 1.2 commit

+++ src/net/jitse/simplefactions/commands/subcommands/UnclaimCommand.java
@@ -11,6 +11,9 @@ import org.bukkit.Chunk;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
+import java.util.ArrayList;
+import java.util.List;
+
 /**
  * Created by Jitse on 21-7-2017.
  */
@@ -33,24 +36,36 @@ public class UnclaimCommand extends SubCommand {
                 player.sendMessage(Chat.format(Settings.NO_CLAIMED_LAND));
                 return;
             }
+            List<Chunk> pending = new ArrayList<>();
             for(Chunk chunk : faction.getClaimedChunks()){
+                List<Home> temp = new ArrayList<>();
                 for(Home home : faction.getHomes()){ // Don't have to loop through ALL factions, because only faction members can place homes on their land!
                     if(chunk.equals(home.getLocation().getChunk())){
-                        faction.getHomes().remove(home);
+                        temp.add(home);
                         SimpleFactions.getInstance().getMySql().execute("DELETE FROM FactionHomes WHERE faction=? AND name=?;", faction.getName(), home.getName());
                     }
                 }
+                for(Home home : temp) faction.getHomes().remove(home);
+                pending.add(chunk);
+            }
+            for(Chunk chunk : pending){
                 faction.unclaimChunk(chunk);
             }
             player.sendMessage(Chat.format(Settings.UNCLAIMED_ALL_LAND));
         } else{
             Chunk chunk = player.getLocation().getChunk();
+            if(!faction.getClaimedChunks().contains(chunk)){
+                player.sendMessage(Chat.format(Settings.INVALID_COMMAND_USAGE));
+                return;
+            }
+            List<Home> temp = new ArrayList<>();
             for(Home home : faction.getHomes()){
                 if(chunk.equals(home.getLocation().getChunk())){
-                    faction.getHomes().remove(home);
+                    temp.add(home);
                     SimpleFactions.getInstance().getMySql().execute("DELETE FROM FactionHomes WHERE faction=? AND name=?;", faction.getName(), home.getName());
                 }
             }
+            for(Home home : temp) faction.getHomes().remove(home);
             faction.unclaimChunk(chunk);
             player.sendMessage(Chat.format(Settings.UNCLAIMED_LAND));
         }
